+++ examples/flask/app/__init__.py
@@ -2,7 +2,7 @@ from flask import Flask
 from flask.ext.sqlalchemy import SQLAlchemy
 
 app = Flask(__name__)
-app.config.from_object('config')
+app.config.from_object("config")
 db = SQLAlchemy(app)
 
-from app import views, models           # noqa: E402,F401
+from app import views, models  # noqa: E402,F401