+++ examples/flask/app/models.py
@@ -4,12 +4,16 @@ from app import db
 
 
 class User(db.Model):
-    __tablename__ = 'users'
+    __tablename__ = "users"
 
     id = db.Column(db.String, nullable=False, primary_key=True)
     created = db.Column(db.DateTime, default=datetime.utcnow, nullable=False)
-    updated = db.Column(db.DateTime, default=datetime.utcnow, nullable=False,
-                        onupdate=datetime.utcnow)
+    updated = db.Column(
+        db.DateTime,
+        default=datetime.utcnow,
+        nullable=False,
+        onupdate=datetime.utcnow,
+    )
     name = db.Column(db.String, nullable=False)
     profile_url = db.Column(db.String, nullable=False)
     access_token = db.Column(db.String, nullable=False)