+++ examples/flask/run.py
@@ -2,4 +2,4 @@
 from app import app, db
 
 db.create_all()
-app.run(host='0.0.0.0', port=8000)
+app.run(host="0.0.0.0", port=8000)